requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
python-levenshtein==0.12.0
scikit-learn==0.17.1
numpy==1.11.1
scipy==0.18.1
nltk==3.2.2
